requirements.txt
Pillow==9.0.1
psycopg2-binary==2.9.3
tk==0.1.0
